+++ game/server/baseentity.h
@@ -1068,7 +1068,7 @@ public:
 	// OLD VERSION! Use the struct version
 	void FireBullets( int cShots, const Vector &vecSrc, const Vector &vecDirShooting, 
 		const Vector &vecSpread, float flDistance, int iAmmoType, int iTracerFreq = 4, 
-		int firingEntID = -1, int attachmentID = -1, int iDamage = 0, 
+		int firingEntID = -1, int attachmentID = -1, float iDamage = 0, 
 		CBaseEntity *pAttacker = NULL, bool bFirstShotAccurate = false, bool bPrimaryAttack = true );
 	virtual void ModifyFireBulletsDamage( CTakeDamageInfo* ) {}
 
@@ -2601,7 +2601,7 @@ inline void CBaseEntity::DecrementTransmitStateOwnedCounter()
 inline void CBaseEntity::FireBullets( int cShots, const Vector &vecSrc, 
 	const Vector &vecDirShooting, const Vector &vecSpread, float flDistance, 
 	int iAmmoType, int iTracerFreq, int firingEntID, int attachmentID,
-	int iDamage, CBaseEntity *pAttacker, bool bFirstShotAccurate, bool bPrimaryAttack )
+	float iDamage, CBaseEntity *pAttacker, bool bFirstShotAccurate, bool bPrimaryAttack )
 {
 	FireBulletsInfo_t info;
 	info.m_iShots = cShots;
